Fix for Bug #5: Flooding not limited to water's edge
Assert wasn't checking if the tile was a tile from the river's edge and was picking at random. Also, the function checking if it was a river was counting values up to WOODS5 which was past the last river edge.

To fix this, I changed the upper bound from WOODS5 to LASTRIVEDGE. Then to fix the error with the assert statement, I simply added an and statement to check that it was part of the river's edge.