Fix for Bug #6: Building underwater wires costs the wrong amount
It had the cost set as 5 with the variable WIRE_COST. Although UNDERWATER_WIRE_COST was initialized, it wasn't really used. Thus, putting it only used the value costing 5.

To fix this, I changed the cost to the value holding 25 in the case that the wire was being placed in the river. I also deleted the one in the channel case since it would reset cost to 5.