Created coordinate class
Plan to store a list of coordinates. Thus, I thought it would be easier to make a class for it rather than another (size 2, x,y) list inside a list.

This class just stores the values of a single coordinate which has properties x, y. It can be returned and set with getters and setters.