Updated pollutionAverage
Goal: keep pollutionAverage either the same or increasing, but no longer decreasing (since at that point pollution can't be fixed) once it hits a certain point.

To fo this, made a boolean variable to see whether the pollution was permanent. Basically the if statement just checks whether to update pollutionAverage. If not it'll stay the same.